Accept FS#7264 'Build with -Os switch for coldfire targets'. Introduces MEM_FUNCTION_WRAPPERS(api) macro which adds wrappers functions to a plugin to make plugins link correctly when gcc calls mem* functions directly.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@13625 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugins/chip8.c
@@ -1062,7 +1062,7 @@ static unsigned char beep[]={255,
 111,181,184,144, 17,148, 21,101,166,227,100, 86, 85, 85, 85}; 
 
 /* callback to request more mp3 data */
-void callback(unsigned char** start, int* size)
+void callback(unsigned char** start, size_t* size)
 {
     *start = beep; /* give it the same frame again */
     *size = sizeof(beep);
