+++ firmware/export/mp3_playback.h
@@ -39,7 +39,7 @@ void demand_irq_enable(bool on);
 /* new functions, exported to plugin API */
 void mp3_play_init(void);
 void mp3_play_data(const unsigned char* start, int size,
-    void (*get_more)(unsigned char** start, int* size) /* callback fn */
+    void (*get_more)(unsigned char** start, size_t* size) /* callback fn */
 );
 void mp3_play_pause(bool play);
 bool mp3_pause_done(void);
